# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AYUNEX Backend",
    "description": "Information, not legal advice. Unreviewed routes do not establish approval, patentability or ABS compliance.",
    "version": "1.1.0"
  },
  "paths": {
    "/api/v1/corpus/catalog": {
      "get": {
        "tags": [
          "Official corpus"
        ],
        "summary": "Catalog",
        "operationId": "catalog_api_v1_corpus_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/corpus/status": {
      "get": {
        "tags": [
          "Official corpus"
        ],
        "summary": "Status",
        "operationId": "status_api_v1_corpus_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/corpus/verify": {
      "post": {
        "tags": [
          "Official corpus"
        ],
        "summary": "Verify",
        "operationId": "verify_api_v1_corpus_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuidanceAnswer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/corpus/documents": {
      "get": {
        "tags": [
          "Official corpus"
        ],
        "summary": "Documents",
        "operationId": "documents_api_v1_corpus_documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/corpus/documents/{document_id}/passages": {
      "get": {
        "tags": [
          "Official corpus"
        ],
        "summary": "Passages",
        "operationId": "passages_api_v1_corpus_documents__document_id__passages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/corpus/passages/{passage_id}/reviews": {
      "get": {
        "tags": [
          "Official corpus"
        ],
        "summary": "Reviews",
        "operationId": "reviews_api_v1_corpus_passages__passage_id__reviews_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "passage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Passage Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/corpus/search": {
      "get": {
        "tags": [
          "Official corpus"
        ],
        "summary": "Find",
        "operationId": "find_api_v1_corpus_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 1000,
              "title": "Q"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Z]{2}$",
              "default": "IN",
              "title": "Country"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "As Of"
            }
          },
          {
            "name": "reviewed_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Reviewed Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/corpus/passages/{passage_id}/review": {
      "patch": {
        "tags": [
          "Official corpus"
        ],
        "summary": "Review",
        "operationId": "review_api_v1_corpus_passages__passage_id__review_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "passage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Passage Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewPassage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Citation": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "authority": {
            "type": "string",
            "title": "Authority"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "pattern": "^https://",
            "title": "Url"
          },
          "provision": {
            "type": "string",
            "minLength": 1,
            "title": "Provision"
          },
          "excerpt": {
            "type": "string",
            "minLength": 1,
            "title": "Excerpt"
          },
          "version": {
            "type": "string",
            "minLength": 1,
            "title": "Version"
          },
          "jurisdiction": {
            "$ref": "#/components/schemas/Jurisdiction"
          },
          "effective_from": {
            "type": "string",
            "format": "date",
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "retrieved_at": {
            "type": "string",
            "title": "Retrieved At"
          },
          "review_status": {
            "type": "string",
            "const": "VERIFIED",
            "title": "Review Status"
          },
          "access_class": {
            "type": "string",
            "const": "PUBLIC",
            "title": "Access Class"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "authority",
          "title",
          "url",
          "provision",
          "excerpt",
          "version",
          "jurisdiction",
          "effective_from",
          "retrieved_at",
          "review_status",
          "access_class"
        ],
        "title": "Citation"
      },
      "Claim": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text"
          },
          "jurisdiction": {
            "$ref": "#/components/schemas/Jurisdiction"
          },
          "citation_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Citation Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "text",
          "jurisdiction",
          "citation_ids"
        ],
        "title": "Claim"
      },
...
      "GuidanceAnswer": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "context_hash": {
            "type": "string",
            "title": "Context Hash"
          },
          "jurisdiction": {
            "$ref": "#/components/schemas/Jurisdiction"
          },
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "support": {
            "type": "string",
            "enum": [
              "SUPPORTED_IN_SCOPE",
              "MISSING_FACTS",
              "MISSING_EVIDENCE",
              "CONFLICTING_SOURCES",
              "OUT_OF_SCOPE",
              "REVIEW_REQUIRED"
            ],
            "title": "Support"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/GuidanceSection"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Sections"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Citations"
          },
          "retrieval_trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalTrace"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "context_hash",
          "jurisdiction",
          "as_of_date",
          "support",
          "sections",
          "citations"
        ],
        "title": "GuidanceAnswer"
      },
      "GuidanceSection": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "support": {
            "type": "string",
            "enum": [
              "SUPPORTED_IN_SCOPE",
              "MISSING_FACTS",
              "MISSING_EVIDENCE",
              "CONFLICTING_SOURCES",
              "OUT_OF_SCOPE",
              "REVIEW_REQUIRED"
            ],
            "title": "Support"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "title": "Claims"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "support",
          "claims",
          "reason"
        ],
        "title": "GuidanceSection"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Jurisdiction": {
        "properties": {
          "layer": {
            "type": "string",
            "enum": [
              "NATIONAL",
              "TREATY_FRAMEWORK",
              "EXPORT_MARKET"
            ],
            "title": "Layer"
          },
          "country": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z]{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "framework": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "TRIPS",
                  "CBD",
                  "NAGOYA",
                  "GRATK",
                  "PCT",
                  "MADRID",
                  "HAGUE",
                  "BUDAPEST"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Framework"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "layer"
        ],
        "title": "Jurisdiction"
      },
...
      "RetrievalTrace": {
        "properties": {
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "mode": {
            "type": "string",
            "enum": [
              "LOCAL_EXTRACTIVE",
              "HOSTED_RAG"
            ],
            "title": "Mode"
          },
          "steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Steps"
          },
          "repair_count": {
            "type": "integer",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Repair Count"
          },
          "translation": {
            "type": "string",
            "const": "NOT_PERFORMED",
            "title": "Translation"
          },
          "source_versions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Versions"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "verification": {
            "type": "string",
            "enum": [
              "NO_CLAIMS",
              "EXACT_QUOTES_VERIFIED",
              "MODEL_REVIEW_PASSED",
              "REJECTED"
            ],
            "title": "Verification"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "engine",
          "mode",
          "steps",
          "repair_count",
          "translation",
          "source_versions",
          "evidence_ids",
          "verification"
        ],
        "title": "RetrievalTrace"
      },
...
      "ReviewPassage": {
        "properties": {
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "expected_sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Expected Sha256"
          },
          "provision": {
            "type": "string",
            "maxLength": 300,
            "minLength": 2,
            "title": "Provision"
          },
          "effective_from": {
            "type": "string",
            "format": "date",
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "review_valid_until": {
            "type": "string",
            "format": "date",
            "title": "Review Valid Until"
          },
          "notes": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 20,
            "title": "Notes"
          },
          "conflict": {
            "type": "boolean",
            "title": "Conflict",
            "default": false
          },
          "expected_review_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Review Revision",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "approved",
          "expected_sha256",
          "provision",
          "effective_from",
          "review_valid_until",
          "notes"
        ],
        "title": "ReviewPassage"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}